Enhance picture tag with alt text and absolute_url
This fixes two problems with `picture`:

1. Images in RSS were broken, because the URL wasn't absolute,
2. We weren't able to provide alt text seperate from the image caption,
   which meant that alt text wasn't very good.

This is done by taking some code from Liquid which isn't supported in
Jekyll yet and adding arguments to tags inspired by Liquid 5.0.0's
`render`: https://shopify.github.io/liquid/tags/template/

We take the `input` value (known as `markup` in Liquid), parse it with
some regular expressions and then tidy it up a bit. We didn't need more
than what's taken here, but the source material supports a lot more than
this.

If the `url:` parameter isn't provided, we raise a syntax error. Jekyll
causes some output here, so we also need to silence that. In future it
might be nice to warn on no `alt` text too.

This commit also updates previous references to the picture tag.

https://github.com/Shopify/liquid/blob/c99c93255d55129652e262e87c27d359628f71e6/lib/liquid/tags/render.rb
https://github.com/Shopify/liquid/blob/c99c93255d55129652e262e87c27d359628f71e6/lib/liquid/expression.rb
http://michaelay.github.io/blog/2014/12/15/suppress-stdout-and-stderr-when-running-rspec/