Fix filter usage in tags
It turns out you can't use a Liquid filter inside a tag, so the
previous commit broke when used.

This pulls in the existing filter and uses it directly — we need to make
the context accessible, but otherwise it's not too bad.

In the mock site used for the tests, the `site_url` isn't set so all we
get is the addition of `/` but that's fine for now.

https://github.com/nickcharlton/site/commit/b9f880b38534243096a523a7c2ef0ec9c9f8d1a4