Add a tags page and link to tags from posts
This takes a simple approach of linking to a single page which has tags
on it and anchors to link to the section. Then, we can link to the tags
from the posts to go there.

We also extract a partial (using Jekyll includes) to make it nicer to
reuse posts rendered in lists.

Inspired by: https://www.jokecamp.com/blog/listing-jekyll-posts-by-tag/